+++ NEWS.md
@@ -4,8 +4,10 @@ version 1.0.7.9000
 =============
 * drop support for R 3.6
 * update Eliot's email address
+* new functions: `speciesInStudyArea` and `species
 * remove undeclared dependency package `Require`
 * age imputation in `makeAndCleanInitialCohortData` can now be turned off.
+* fix bug in `LANDISDisp()`: skip dispersal when src or rcv data.tables are empty
 
 version 1.0.7
 =============